Finished Kid Mode (I think)
Kid Mode will now be automatically enabled on the 1st of April,
and SHOULD be available for the user to change if they want to
once they have cleared the ALL LEVELS campaign, but, since my
locally compiled version cannot play Tokyo or Osaka anymore
without segfaulting, I can't actually test this unless it is
compiled on one of RWS's magic computers which make it work
somehow. I will keep working on getting that fixed and then,
if it turns out that the unlock actually works as intended,
I'll merge testing to my default & make a pull request.