Generate MultiAlpha tables on the fly
I don't know why the cutscene MLP files exist. The game can generate
the same data on the fly in a fraction of a second. That's what I've
made it do - you can now delete all the MLP files from the cutscene
directory and remove the Alpha lines from postal_plus_realms.ini,
and the game will still work exactly the same. The advantage, apart
from saving, like, 2 MB of space, is that new cutscenes can be made
just by copying and editing an existing BMP file, without the need
to work with a difficult proprietary format. This should help out
with adding The Carnival to the game, although obviously, the actual
hood format is a more pressing concern for that.