White Fuzz Fix
Doesn't look like the problem was in my code after all. For some
reason the output palette is different to the input palette beyond
index 0 being transparent, and when GIMP tries to map the Sprys
to the base image's palette, it introduces the white fuzz. I added
a button to convert a BMP to PNG using RSPiX to get the palette, which
solved the issue. This is worth remembering though because it may
cause trouble re-adding the images. Fingers crossed it won't.