Removed use of NamedTemporaryFile
pImageToRImage() previously used a NamedTemporaryFile.
NamedTemporaryFile created an FS node, and therefore a name, for
a created temporary file. I think the actual data in the file was
still stored in memory, but this meant that access had to go through
the file system, which is slow and ugly. I changed it to use a regular
TemporaryFile. The PImage writes to it, the data then gets read into
a string, that string is written into an RFile elsewhere in memory,
the RFile is sought (seeked?) to the beginning, and the RImage loads
from it. This means everything happens in memory, so it's noticeably
faster, and it's a lot nicer.